Solve the bug, when the DC-DC does not work normally
DCSolver.cpp :
reduce the search range, i.e. the lower bound in function
SolveItGivenDCInput.
We found sometimes, during charge process, the equation is
solved, but the current is negative, i.e., current drawn from
ees bank, which is not correct. so we set the lower bound for
searching the solution be the bank oc voltage

main.cpp :
Remember to set the configuration first during the initilizing
the ees bank, then set the charge